Fix missing functions
Five linker issues resolved with a new **`src/compat.c`** file plus small fixes elsewhere:

| Symbol | Root cause | Fix |
| --- | --- | --- |
| `memmove` | SGDK's `libmd.a` has `memcpy` but not `memmove` (needed for safe struct copies in `SaveGame`/`LoadGame`) | Implemented in `compat.c` |
| `snprintf` / `vsnprintf` | SGDK has `sprintf` but not the size-limited variants | Implemented in `compat.c` using SGDK's `vsprintf` internally |
| `sqrt` | Not in `libmd.a` or `libgcc.a` with `-nostdlib` | Integer sqrt implemented in `compat.c`; cast added in `Encounter.c` |
| `SRAM_readBuffer` / `SRAM_writeBuffer` | SGDK 1.70 only has byte/word primitives (`SRAM_readByte`, `SRAM_writeByte`) --- the buffer API was added in later SGDK versions | Implemented in `compat.c` using `SRAM_readByte`/`SRAM_writeByte` |
| `MenuEraseStatus` | Palm OS menu helper never stubbed | No-op stub in `compat.c` |
| `main()` signature | SGDK 1.70 expects `bool main()` (LTO type check fails on `int main(void)`) | Changed to `bool main(void)` returning `FALSE` |

+++ Src/Encounter.c
@@ -1382,7 +1382,7 @@ static void DrawEncounterForm()
 		DrawChars( SBuf, 6, 88 );
 	}			
 
-	d = sqrt( Ship.Tribbles/250 );
+	d = sqrt( (int)(Ship.Tribbles/250) );
 	for (i=0; i<d; ++i)
 	{
 		objindex = FrmGetObjectIndex( frmP, EncounterTribble0Button +
